Always bind against a consistent FileAttributes enum
My previous "clever" substitution of FileAttributes when the portable profile didn't offer it meant that a library that compiles against the portable PInvoke library would fail when run on desktop at runtime because a type would be missing. We must ensure that the portable library is a subset of the desktop library. This corrects that.