Throw exception based on NTStatus->HRESULT instead of on NTStatus->Win32ErrorCode
Converting NTStatus to Win32 error codes is incomplete, and the recommendation generally seems to be to convert to HRESULT instead.

The impetus for this change is that in throwing exceptions for NTStatus errors, I was getting exceptions for the inability to convert the particular NTStatus value to an equivalent Win32 error. So in general the approach of throwing Win32Exception for NTStatus returning methods seems faulty. This change switches it to throw using Marshal.ThrowExceptionForHR instead after converting the NTStatus to an HRESULT, which is more straightforward.