Fix code generator's treatment of directives
We should take all trivia up to the nearest directive.

As it was, we were taking all trivia except directives, which means we were dropping #if/#endif but picking up the content between them when it was 'inactive' making it also trivia, till we excluded the directives and then we activated the code accidentally.

This was exposed when we updated to the latest codegeneration.roslyn package in order to fix some other directive handling.